// frontend/lib/api/models/EECombinationUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Mise à jour d'une combinaison EE.
 */
export type EECombinationUpdate = {
    title?: (string | null);
    order?: (number | null);
    task1_instruction?: (string | null);
    task1_word_min?: (number | null);
    task1_word_max?: (number | null);
    task2_instruction?: (string | null);
    task2_word_min?: (number | null);
    task2_word_max?: (number | null);
    task3_title?: (string | null);
    task3_document_1?: (string | null);
    task3_document_2?: (string | null);
    task3_word_min?: (number | null);
    task3_word_max?: (number | null);
};